# Bring over OneElement for scalar getindex