I think this is maybe is a precondition for https://github.com/FluxML/Zygote.jl/pull/1328
so that we know that deleting the Zygote rules will not cause performance regressions.

Still WIP as i chase down failing tests.