@oxinabox 2023-05-19
We should think about improving inferability so it stops making unionings over wether or not it is going to `reshape`.
One part i think is making the axes at type-param here.
But i think we also need changes in CRC


```
"""
    OneElement(val, ind, axes) <: AbstractArray

Extremely simple `struct` used for the gradient of scalar `getindex`.
"""
struct OneElement{T,N,I,A} <: AbstractArray{T,N}
  val::T
  ind::I
  OneElement(val::T, ind::I, axes) where {T<:Number, I<:NTuple{N,Int}, A<:NTuple{N,AbstractUnitRange}} where {N} = new{T,N,I, axes}(val, ind)
end
Base.size(::OneElement{<:Any,<:Any,<:Any, A}) where A = map(length, A)
Base.axes(::OneElement{<:Any,<:Any,<:Any, A}) where A = A
Base.getindex(xs::OneElement{T,N}, i::Vararg{Int,N}) where {T,N} = ifelse(i==xs.ind, xs.val, zero(T))
```

Still this is a small union and doesn't really matter
but that is why i have to drop a lot of the inference tests