@oxinabox 2023-05-22
Before
```julia
julia> x = rand(300,300);

julia> dx = OneElement(3.14, (231, 21), axes(x));

julia> @btime $x + $(collect(dx));  # Baseline
  31.978 μs (2 allocations: 703.17 KiB)

julia> @btime $x + $dx;
  61.043 μs (2 allocations: 703.17 KiB)

```

After:
```
julia> @btime $x + $(dx);
  21.056 μs (2 allocations: 703.17 KiB)
```

The orig version of this PR (and thus more or less what is in Zygote` was 2x slower than baseline, for accumulation,
with my fix it is now 30% faster than baseline.
(for a 300x300 matrix. This gets better the bigger the matrix).

For allocation no change vs what in Zygote, it remains 77,000x faster than baseline
```julia
julia> @btime collect(OneElement(3.14, (231, 21), axes($x)));  # approximate baseline
  54.806 μs (2 allocations: 703.17 KiB)

julia> @btime OneElement(3.14, (231, 21), axes($x));
  1.748 ns (0 allocations: 0 bytes)
```

I think this is worth a the type instability of going to a small union