@mcabbott 2023-05-25
One of the things I don't like about OneElement is that it's tied to Array. Somewhere I had a prototype which stored BroadcastStyle too, so that it remembers that you indexed e.g. a CuArray. 

Maybe worth mentioning that https://github.com/JuliaArrays/FillArrays.jl/pull/235 added essentially the same struct that Zygote has to FillArrays. CR could consider getting it from there, although it's so simple that owning it is fine. I think FillArrays may overload some `mul!` methods etc, maybe worth thinking about whether any of those are desirable for CR's purposes.

Sorry not really a review, a bit swamped.