@ToucheSir 2023-05-26
I have nothing intelligent to add on top of what Michael said, but it's cool that FillArrays has `OneElement` now. Would a port of the benchmark in https://github.com/FluxML/Zygote.jl/pull/962 help with making a decision here? I can look into that if there's interest.