@oxinabox 2023-06-01
FillArrays is still pending in https://github.com/JuliaDiff/ChainRules.jl/issues/46

I am happy merging this as is, with other stuff to come in follow-ups.
but I do need one of you to to approve.